requirements.txt
click==7.1.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyMySQL==0.9.3
PyYAML==5.3.1
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
